@@ -2,6 +2,6 @@
 
 declare(strict_types = 1);
 
-test('that true is true', function () {
+test('that true is true', function (): void {
     expect(true)->toBeTrue();
 });
